General return value of size - methods is now size_t instead of ptrdiff_t. Changed Snowfall from boolean decision of snow / no snow to a snowfraction. This includes sleet and can resolve numerical when T changes slightly below zero

+++ cmf/cmf_core_src/math/integrators/integrator.h
@@ -99,9 +99,9 @@ namespace cmf {
 			bool use_OpenMP;
 
 			/// returns the number of state variables
-			ptrdiff_t size() const
+			size_t size() const
 			{
-				return (ptrdiff_t)m_States.size();
+				return m_States.size();
 			}
 			/// Returns the statevariable at position
 			/// Simplifies the assessment of state variables
